fix(lhr): fix pool.status zero display for sub-1 difficulty miners
pool_stats_t stored diff shares, rejects, and best_diff as int64_t.
With lhr difficulty values (e.g. 0.001), every accumulated share truncated
to 0, causing cascading failures in pool.status output:
- hashrate fields showed "0" (dsps stayed 0 → ghs = 0)
- accepted/rejected showed 0
- bestshare showed 0

Fix all affected sites:

stratifier.c:
- pool_stats_t: change unaccounted_diff_shares, accounted_diff_shares,
  unaccounted_rejects, accounted_rejects, best_diff from int64_t to double
- stats accumulator loop: promote local unaccounted_diff_shares and
  unaccounted_rejects to double (unaccounted_shares stays int64_t)
- pool.status JSON_CPACK: change sI→sf for accepted/rejected/bestshare;
  raise JSON_REAL_PRECISION from 3 to 6
- stratifier_stats API JSON_CPACK: change sI→sf for accepted/rejected
- pool restart reader: json_get_int64→json_get_double for accepted,
  rejected, bestshare

libckpool.c:
- suffix_string: fix non-decimal branch to use %.3g instead of %d cast,
  so sub-1 H/s values (e.g. 0.047 H/s) display correctly instead of "0"

+++ src/libckpool.c
@@ -2102,7 +2102,7 @@ void suffix_string(double val, char *buf, size_t bufsiz, int sigdigits)
 		if (decimal)
 			snprintf(buf, bufsiz, "%.3g%s", dval, suffix);
 		else
-			snprintf(buf, bufsiz, "%d%s", (unsigned int)dval, suffix);
+			snprintf(buf, bufsiz, "%.3g%s", dval, suffix);
 	} else {
 		/* Always show sigdigits + 1, padded on right with zeroes
 		 * followed by suffix */
